refactor(Link): move the prefetch helpers into the link utils
